backend/requirements.txt
# Multi-tenant Customer Support AI — Backend
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Auth
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# OpenAI & LangChain
openai>=1.55.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.2.0
langchain-text-splitters>=0.3.0

# Vector DB
chromadb>=0.5.0

# Streaming
sse-starlette>=2.0.0

# Payments (Stripe integration ready)
stripe>=10.0.0

# Pydantic (EmailStr)
email-validator>=2.0.0
